% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-plots.R
\name{app-plots}
\alias{app-plots}
\alias{.plot_theme}
\alias{.plot_title_style}
\alias{.build_plot}
\alias{.build_data}
\title{App Outputs}
\usage{
.plot_theme()

.plot_title_style(ptitle, psubtitle)

.build_plot(pdata, bin_count)

.build_data(data)
}
\arguments{
\item{ptitle}{plot title}

\item{psubtitle}{plot subtitle}

\item{pdata}{plot data}

\item{bin_count}{total count of bins used in plot}

\item{data}{data to build plot data from}
}
\description{
Functions to generate and prepare outputs presented by the package app.
}
\section{Functions}{
\itemize{
\item \code{.plot_theme()}: provides the base theme for plots

\item \code{.plot_title_style()}: Style the plot title/subtitle

\item \code{.build_plot()}: diagnostic plot

\item \code{.build_data()}: diagnostic plot

}}
