% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-run.R
\name{app-run}
\alias{app-run}
\alias{apptemplate1}
\title{Run App}
\usage{
apptemplate1(...)
}
\arguments{
\item{...}{named args to pass to shinyApp's options parameter}
}
\description{
Functions to run the package shiny app
}
\section{Functions}{
\itemize{
\item \code{apptemplate1()}: returns app object for subsequent execution

}}
