% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-server.R
\name{app-server}
\alias{app-server}
\alias{app_server}
\title{App Server Function}
\usage{
app_server()
}
\description{
App Server Function
}
\section{Functions}{
\itemize{
\item \code{app_server()}: server function for app

}}
