% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-ui.R
\name{app-ui}
\alias{app-ui}
\alias{app_ui}
\title{Ui elements of Template App}
\usage{
app_ui()
}
\description{
Ui elements of Template App
}
\section{Functions}{
\itemize{
\item \code{app_ui()}: UI function for app

}}
