% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apptemplate1-package.R
\docType{package}
\name{apptemplate1-package}
\alias{apptemplate1}
\alias{apptemplate1-package}
\title{apptemplate1: Template Single-App Package}
\description{
Template for a shiny app built as R package, with full coverage/CI unit testing of App and package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-data-science/apptemplate1}
  \item Report bugs at \url{https://github.com/r-data-science/apptemplate1/issues}
}

}
\author{
\strong{Maintainer}: Bobby Fatemi \email{bfatemi07@gmail.com}

}
\keyword{internal}
