% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-utils.R
\name{app-utils}
\alias{app-utils}
\alias{new_waiter}
\alias{waiter_html}
\alias{generate_report}
\alias{is_ci}
\alias{is_testing}
\alias{get_app_colors}
\alias{get_session_dir}
\alias{create_session_dir}
\alias{clear_session_dir}
\title{App Utils}
\usage{
new_waiter(session = NULL)

waiter_html(msg)

generate_report(file = "report.html")

is_ci()

is_testing()

get_app_colors()

get_session_dir(error = FALSE)

create_session_dir()

clear_session_dir()
}
\arguments{
\item{session}{shiny session object}

\item{msg}{message for waiter screen}

\item{file}{file name for report download passed from app}

\item{error}{When session NF, FALSE (default) returns NULL, else throw error}
}
\description{
Functions required to execute and facililate an application user session.
}
\section{Functions}{
\itemize{
\item \code{new_waiter()}: create a new waiter object

\item \code{waiter_html()}: get html for waiter progress page

\item \code{generate_report()}: Generate report and return download link

\item \code{is_ci()}: returns TRUE if called on CI

\item \code{is_testing()}: returns TRUE if called while testing

\item \code{get_app_colors()}: Set Plot colors

\item \code{get_session_dir()}: creates and returns app dir path

\item \code{create_session_dir()}: Run and Publishing Functions

\item \code{clear_session_dir()}: Run and Publishing Functions

}}
